requirements.txt
Django==3.1
psycopg2-binary==2.8.6
Pillow==7.1
django-filter